+++ Mailer/extensions/mailer-module/src/migrations/20210524145621_add_mail_template_id_to_mail_types.php
@@ -1,0 +1,15 @@
+<?php
+declare(strict_types=1);
+
+use Phinx\Migration\AbstractMigration;
+
+final class AddMailTemplateIdToMailTypes extends AbstractMigration
+{
+    public function change(): void
+    {
+        $this->table('mail_types')
+            ->addColumn('subscribe_mail_template_id', 'integer', ['default' => null, 'null' => true, 'after' => 'mailer_alias'])
+            ->addForeignKey('subscribe_mail_template_id', 'mail_templates', 'id')
+            ->update();
+    }
+}